Update README.md
Reflect on some of the excellent solutions posted to the APL AoC Wiki, by far more advanced APL programmers. which tells me to think of more parallel solutions. But hopefully, as more puzzles are solved the solutions might move in a different direction. Yet, the solutions written here, do provide the correct answers. And are probably more efficient than when solved in many other languages.